Fix strict route exports
