+++ app/projects/form.tsx
@@ -1,28 +1,52 @@
 import { useMemo } from 'react'
-import { FaBrandsDiscord, FaBrandsGithub } from '~/components/icons'
-import type { MetaFunction } from '@remix-run/node'
 import { Link, useMatches, useNavigate, useParams } from '@remix-run/react'
-import { seo } from '~/utils/seo'
-import type { DocsConfig } from '~/components/Docs'
-import { Docs } from '~/components/Docs'
-import { QueryGGBanner } from '~/components/QueryGGBanner'
-import {
-  availableVersions,
-  latestVersion,
-  repo,
-  useReactQueryDocsConfig,
-  gradientText,
-} from '~/projects/query'
-import type { MenuItem } from '~/projects/query'
+import { generatePath, useMatchesData } from '~/utils/utils'
 import reactLogo from '~/images/react-logo.svg'
 import solidLogo from '~/images/solid-logo.svg'
 import vueLogo from '~/images/vue-logo.svg'
 import svelteLogo from '~/images/svelte-logo.svg'
 import angularLogo from '~/images/angular-logo.svg'
+import { FaBrandsDiscord, FaBrandsGithub } from '~/components/icons'
 import type { AvailableOptions } from '~/components/Select'
-import { generatePath } from '~/utils/utils'
+import type { ReactNode } from 'react'
 
-const frameworks = {
+export type FrameworkMenu = {
+  framework: string
+  menuItems: MenuItem[]
+}
+
+export type MenuItem = {
+  label: string | ReactNode
+  children: {
+    label: string | ReactNode
+    to: string
+  }[]
+}
+
+export type GithubDocsConfig = {
+  docSearch: {
+    appId: string
+    apiKey: string
+    indexName: string
+  }
+  menu: MenuItem[]
+  frameworkMenus: FrameworkMenu[]
+}
+
+export type Framework = 'react' | 'svelte' | 'vue' | 'solid'
+
+//
+
+export const repo = 'tanstack/form'
+
+export const latestBranch = 'main'
+export const latestVersion = 'v0'
+export const availableVersions = ['v0']
+
+export const gradientText =
+  'inline-block text-transparent bg-clip-text bg-gradient-to-r from-yellow-500 to-yellow-600'
+
+export const frameworks = {
   react: { label: 'React', logo: reactLogo, value: 'react' },
   solid: { label: 'Solid', logo: solidLogo, value: 'solid' },
   vue: { label: 'Vue', logo: vueLogo, value: 'vue' },
@@ -30,13 +54,13 @@ const frameworks = {
   angular: { label: 'Angular', logo: angularLogo, value: 'angular' },
 }
 
-const logo = (version?: string) => (
+export const createLogo = (version?: string) => (
   <>
     <Link to="/" className="font-light">
       TanStack
     </Link>
-    <Link to=".." className={`font-bold`}>
-      <span className={`${gradientText}`}>Query</span>{' '}
+    <Link to="../../" className={`font-bold`}>
+      <span className={`${gradientText}`}>Form</span>{' '}
       <span className="text-sm align-super">
         {version === 'latest' ? latestVersion : version}
       </span>
@@ -44,7 +68,7 @@ const logo = (version?: string) => (
   </>
 )
 
-const localMenu: MenuItem = {
+export const localMenu: MenuItem = {
   label: 'Menu',
   children: [
     {
@@ -70,33 +94,31 @@ const localMenu: MenuItem = {
   ],
 }
 
-export const meta: MetaFunction = () => {
-  return seo({
-    title:
-      'TanStack Query Docs | React Query, Solid Query, Svelte Query, Vue Query',
-  })
+export function getBranch(argVersion?: string) {
+  const version = argVersion || latestVersion
+
+  return ['latest', latestVersion].includes(version) ? latestBranch : version
 }
 
-export default function RouteFrameworkParam() {
+//
+
+export const useReactFormDocsConfig = () => {
   const matches = useMatches()
   const match = matches[matches.length - 1]
-  const navigate = useNavigate()
   const params = useParams()
-  const framework = params.framework
-  const version = params.version
-  let config = useReactQueryDocsConfig(version)
+  const version = params.version!
+  const framework =
+    params.framework || localStorage.getItem('framework') || 'react'
+  const navigate = useNavigate()
 
-  const docsConfig = useMemo(() => {
-    const frameworkMenu = config.menu.find((d) => d.framework === framework)
-    if (!frameworkMenu) return null
-    return {
-      ...config,
-      menu: [localMenu, ...(frameworkMenu?.menuItems || [])],
-    } as DocsConfig
-  }, [framework, config])
+  const config = useMatchesData(`/form/${version}`) as GithubDocsConfig
+
+  const frameworkMenuItems =
+    config.frameworkMenus.find((d) => d.framework === framework)?.menuItems ??
+    []
 
   const frameworkConfig = useMemo(() => {
-    const availableFrameworks = config.menu.reduce(
+    const availableFrameworks = config.frameworkMenus.reduce(
       (acc: AvailableOptions, menuEntry) => {
         acc[menuEntry.framework as string] =
           frameworks[menuEntry.framework as keyof typeof frameworks]
@@ -117,14 +139,14 @@ export default function RouteFrameworkParam() {
         navigate(url)
       },
     }
-  }, [config.menu, framework, match, navigate])
+  }, [config.frameworkMenus, framework, match, navigate])
 
   const versionConfig = useMemo(() => {
     const available = availableVersions.reduce(
       (acc: AvailableOptions, version) => {
-        acc[version.name] = {
-          label: version.name,
-          value: version.name,
+        acc[version] = {
+          label: version,
+          value: version,
         }
         return acc
       },
@@ -150,20 +172,26 @@ export default function RouteFrameworkParam() {
     }
   }, [version, match, navigate])
 
-  return (
-    <>
-      <QueryGGBanner />
-      <Docs
-        {...{
-          logo: logo(version),
-          colorFrom: 'from-rose-500',
-          colorTo: 'to-violet-500',
-          textColor: 'text-violet-500',
-          config: docsConfig!,
-          framework: frameworkConfig,
-          version: versionConfig,
-        }}
-      />
-    </>
-  )
+  return {
+    ...config,
+    menu: [
+      localMenu,
+      // Merge the two menus together based on their group labels
+      ...config.menu.map((d) => {
+        const match = frameworkMenuItems.find((d2) => d2.label === d.label)
+        return {
+          label: d.label,
+          children: [
+            ...d.children.map((d) => ({ ...d, badge: 'core' })),
+            ...(match?.children ?? []).map((d) => ({ ...d, badge: framework })),
+          ],
+        }
+      }),
+      ...frameworkMenuItems.filter(
+        (d) => !config.menu.find((dd) => dd.label === d.label)
+      ),
+    ].filter(Boolean),
+    frameworkConfig,
+    versionConfig,
+  }
 }
