+++ app/projects/router.ts
@@ -1,0 +1,10 @@
+import type { DocsConfig } from '~/components/Docs'
+import { useMatchesData } from '~/utils/utils'
+
+export const v1branch = 'main'
+
+export const useRouterV1Config = () =>
+  useMatchesData('/router/v1') as DocsConfig
+
+export const gradientText =
+  'inline-block text-transparent bg-clip-text bg-gradient-to-r from-lime-500 to-emerald-500'
