+++ app/routes/form.$version.docs.tsx
@@ -1,6 +1,6 @@
 import { Outlet, useParams } from '@remix-run/react'
 import { Docs } from '~/components/Docs'
-import { createLogo, useReactFormDocsConfig } from '~/routes/form'
+import { createLogo, useReactFormDocsConfig } from '~/projects/form'
 
 export default function Component() {
   const { version } = useParams()
