+++ app/routes/query.$version.docs.$framework.$.tsx
@@ -5,7 +5,7 @@ import {
   fetchRepoFile,
   markdownToMdx,
 } from '~/utils/documents.server'
-import { repo, getBranch } from '~/routes/query'
+import { repo, getBranch } from '~/projects/query'
 import { DefaultErrorBoundary } from '~/components/DefaultErrorBoundary'
 import { seo } from '~/utils/seo'
 import removeMarkdown from 'remove-markdown'
