+++ app/routes/query.tsx
@@ -2,66 +2,7 @@ import { Outlet } from '@remix-run/react'
 import type { LoaderFunctionArgs, MetaFunction } from '@remix-run/node'
 import { redirect } from '@remix-run/node'
 import { seo } from '~/utils/seo'
-import { useMatchesData } from '~/utils/utils'
 import { Scarf } from '~/components/Scarf'
-import type { ReactNode } from 'react'
-
-export const repo = 'tanstack/query'
-
-const latestBranch = 'main'
-export const latestVersion = 'v5'
-export const availableVersions = [
-  {
-    name: 'v5',
-    branch: latestBranch,
-  },
-  {
-    name: 'v4',
-    branch: 'v4',
-  },
-  {
-    name: 'v3',
-    branch: 'v3',
-  },
-] as const
-
-export function getBranch(argVersion?: string) {
-  const version = argVersion || latestVersion
-
-  if (version === 'latest') {
-    return latestBranch
-  }
-
-  return (
-    availableVersions.find((v) => v.name === version)?.branch ?? latestBranch
-  )
-}
-
-export type Menu = {
-  framework: string
-  menuItems: MenuItem[]
-}
-
-export type MenuItem = {
-  label: string | ReactNode
-  children: {
-    label: string | ReactNode
-    to: string
-  }[]
-}
-
-export type GithubDocsConfig = {
-  docSearch: {
-    appId: string
-    apiKey: string
-    indexName: string
-  }
-  menu: Menu[]
-  users: string[]
-}
-
-export const useReactQueryDocsConfig = (version?: string) =>
-  useMatchesData(`/query/${version}`) as GithubDocsConfig
 
 export const meta: MetaFunction = () => {
   return seo({
