+++ app/routes/ranger.v1.docs.$.tsx
@@ -6,7 +6,7 @@ import {
   fetchRepoFile,
   markdownToMdx,
 } from '~/utils/documents.server'
-import { v1branch } from '~/routes/ranger.v1'
+import { v1branch } from '~/projects/ranger'
 import { DefaultErrorBoundary } from '~/components/DefaultErrorBoundary'
 import { seo } from '~/utils/seo'
 import removeMarkdown from 'remove-markdown'
