+++ app/routes/router.v1.docs.$.tsx
@@ -6,7 +6,7 @@ import {
   fetchRepoFile,
   markdownToMdx,
 } from '~/utils/documents.server'
-import { v1branch } from '~/routes/router.v1'
+import { v1branch } from '~/projects/router'
 import { DefaultErrorBoundary } from '~/components/DefaultErrorBoundary'
 import { seo } from '~/utils/seo'
 import removeMarkdown from 'remove-markdown'
