+++ app/routes/router.v1.docs.tsx
@@ -2,8 +2,7 @@ import { useMemo } from 'react'
 import { FaBrandsDiscord, FaBrandsGithub } from '~/components/icons'
 import { Link } from '@remix-run/react'
 import type { MetaFunction } from '@remix-run/node'
-import { useRouterV1Config } from '~/routes/router.v1'
-import { gradientText } from '~/routes/router.v1._index'
+import { useRouterV1Config, gradientText } from '~/projects/router'
 import { seo } from '~/utils/seo'
 import type { DocsConfig } from '~/components/Docs'
 import { Docs } from '~/components/Docs'
