+++ app/routes/table.v8.docs.$.tsx
@@ -7,7 +7,7 @@ import {
   fetchRepoFile,
   markdownToMdx,
 } from '~/utils/documents.server'
-import { v8branch } from '~/routes/table.v8'
+import { v8branch } from '~/projects/table'
 import { DefaultErrorBoundary } from '~/components/DefaultErrorBoundary'
 import { seo } from '~/utils/seo'
 import removeMarkdown from 'remove-markdown'
