+++ app/routes/table.v8.docs.tsx
@@ -1,8 +1,7 @@
 import { useMemo } from 'react'
 import { FaBrandsDiscord, FaBrandsGithub } from '~/components/icons'
 import { Link } from '@remix-run/react'
-import { useReactTableV8Config } from '~/routes/table.v8'
-import { gradientText } from '~/routes/table.v8._index'
+import { useReactTableV8Config, gradientText } from '~/projects/table'
 import { seo } from '~/utils/seo'
 import { Docs } from '~/components/Docs'
 import { DefaultErrorBoundary } from '~/components/DefaultErrorBoundary'
