+++ app/routes/table.v8.tsx
@@ -1,11 +1,7 @@
 import { Link, Outlet, useLocation, useSearchParams } from '@remix-run/react'
 import { json } from '@remix-run/node'
 import { DefaultErrorBoundary } from '~/components/DefaultErrorBoundary'
-import type { DocsConfig } from '~/components/Docs'
 import { fetchRepoFile } from '~/utils/documents.server'
-import { useMatchesData } from '~/utils/utils'
-
-export const v8branch = 'main'
 
 export const loader = async () => {
   const config = await fetchRepoFile(
@@ -24,9 +20,6 @@ export const loader = async () => {
 }
 
 export const ErrorBoundary = DefaultErrorBoundary
-
-export const useReactTableV8Config = () =>
-  useMatchesData('/table/v8') as DocsConfig
 
 export default function RouteReactTable() {
   const [params] = useSearchParams()
