+++ app/routes/virtual.v3.docs.tsx
@@ -1,8 +1,7 @@
 import { useMemo } from 'react'
 import { FaBrandsDiscord, FaBrandsGithub } from '~/components/icons'
 import { Link } from '@remix-run/react'
-import { useVirtualV3Config } from '~/routes/virtual.v3'
-import { gradientText } from '~/routes/virtual.v3._index'
+import { useVirtualV3Config, gradientText } from '~/projects/virtual'
 import { seo } from '~/utils/seo'
 import { Docs } from '~/components/Docs'
 import type { MetaFunction } from '@remix-run/react'
